- Fix sign error in Cursor::setSelection - Add debugging input - If cursor::dispatch does not handle the func, then it should not request an update - Dump metrics cache after redraw
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@15443 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/cursor.h
@@ -74,7 +74,7 @@ public:
 	///
 	void setSelection();
 	/// set selection at given position
-	void setSelection(DocIterator const & where, size_t n);
+	void setSelection(DocIterator const & where, int n);
 	///
 	void clearSelection();
 	/// access start of selection
